Adjusting, cleaning and modularizing the code
