+++ dev_daniel/scripts/extract_info_univerity_moron_n_university_rio_cuarto.sql
